+++ src/types/observablemap.ts
@@ -25,69 +25,54 @@ import { declareIterator, Iterator, iteratorSymbol } from "../utils/iterable"
 import { transaction } from "../api/transaction"
 import { referenceEnhancer } from "./modifiers"
 
-/**
- * Map as defined by Typescript's lib.es2015.collection.d.ts
- *
- * Imported here to not require consumers to have these libs enabled in their tsconfig if not actually using maps
- */
-export interface IMap<K, V> {
-    clear(): void
-    delete(key: K): boolean
-    forEach(callbackfn: (value: V, index: K, map: IMap<K, V>) => void, thisArg?: any): void
-    get(key: K): V | undefined
-    has(key: K): boolean
-    set(key: K, value?: V): this
-    readonly size: number
-}
-
 export interface IKeyValueMap<V> {
     [key: string]: V
 }
 
-export type IMapEntry<V> = [string, V]
-export type IMapEntries<V> = IMapEntry<V>[]
+export type IMapEntry<K, V> = [K, V]
+export type IMapEntries<K, V> = IMapEntry<K, V>[]
 
-// In 3.0, change to IObjectMapChange
-export type IMapChange<T> = IMapChangeUpdate<T> | IMapChangeAdd<T> | IMapChangeDelete<T>
+// TODO: rename In 3.0, change to IObjectMapChange
+export type IMapChange<K, V> = IMapChangeUpdate<K, V> | IMapChangeAdd<K, V> | IMapChangeDelete<K, V>
 
-export interface IMapChangeBase<T> {
-    object: ObservableMap<T>
-    name: string
+export interface IMapChangeBase<K, V> {
+    object: ObservableMap<K, V>
+    name: K // actual the key or index, but this is based on the ancient .observe proposal for consistency
 }
 
-export interface IMapChangeUpdate<T> extends IMapChangeBase<T> {
+export interface IMapChangeUpdate<K, V> extends IMapChangeBase<K, V> {
     type: "update"
-    newValue: T
-    oldValue: T
+    newValue: V
+    oldValue: V
 }
 
-export interface IMapChangeAdd<T> extends IMapChangeBase<T> {
+export interface IMapChangeAdd<K, V> extends IMapChangeBase<K, V> {
     type: "add"
-    newValue: T
+    newValue: V
 }
 
-export interface IMapChangeDelete<T> extends IMapChangeBase<T> {
+export interface IMapChangeDelete<K, V> extends IMapChangeBase<K, V> {
     type: "delete"
-    oldValue: T
+    oldValue: V
 }
 
-export interface IMapWillChange<T> {
-    object: ObservableMap<T>
+export interface IMapWillChange<K, V> {
+    object: ObservableMap<K, V>
     type: "update" | "add" | "delete"
-    name: string
-    newValue?: T
+    name: K
+    newValue?: V
 }
 
 const ObservableMapMarker = {}
 
-export type IObservableMapInitialValues<V> = IMapEntries<V> | IKeyValueMap<V> | IMap<string, V>
+export type IObservableMapInitialValues<K, V> = IMapEntries<K, V> | IKeyValueMap<V> | Map<K, V>
 
-export class ObservableMap<V>
-    implements IInterceptable<IMapWillChange<V>>, IListenable, IMap<string, V> {
+export class ObservableMap<K, V>
+    implements Map<K, V>, IInterceptable<IMapWillChange<K, V>>, IListenable {
     $mobx = ObservableMapMarker
-    private _data: { [key: string]: ObservableValue<V | undefined> } = Object.create(null)
-    private _hasMap: { [key: string]: ObservableValue<boolean> } = Object.create(null) // hasMap, not hashMap >-).
-    private _keys: IObservableArray<string> = <any>new ObservableArray(
+    private _data: Map<K, ObservableValue<V>>
+    private _hasMap: Map<K, ObservableValue<boolean>> // hasMap, not hashMap >-).
+    private _keys: IObservableArray<K> = <any>new ObservableArray(
         undefined,
         referenceEnhancer,
         `${this.name}.keys()`,
@@ -98,37 +83,40 @@ export class ObservableMap<V>
     dehancer: any
 
     constructor(
-        initialData?: IObservableMapInitialValues<V>,
+        initialData?: IObservableMapInitialValues<K, V>,
         public enhancer: IEnhancer<V> = deepEnhancer,
         public name = "ObservableMap@" + getNextId()
     ) {
+        if (typeof Map !== "function") {
+            throw new Error(
+                "mobx.map requires Map polyfill for the current browser. Check babel-polyfill or core-js/es6/map.js"
+            )
+        }
+        this._data = new Map()
+        this._hasMap = new Map()
         this.merge(initialData)
     }
 
-    private _has(key: string): boolean {
-        return typeof this._data[key] !== "undefined"
+    private _has(key: K): boolean {
+        return this._data.has(key)
     }
 
-    has(key: string): boolean {
-        if (!this.isValidKey(key)) return false
-        key = "" + key
-        if (this._hasMap[key]) return this._hasMap[key].get()
+    has(key: K): boolean {
+        if (this._hasMap.has(key)) return this._hasMap.get(key)!.get()
         return this._updateHasMapEntry(key, false).get()
     }
 
-    set(key: string, value?: V | undefined) {
-        this.assertValidKey(key)
-        key = "" + key
+    set(key: K, value: V) {
         const hasKey = this._has(key)
         if (hasInterceptors(this)) {
-            const change = interceptChange<IMapWillChange<V>>(this, {
+            const change = interceptChange<IMapWillChange<K, V>>(this, {
                 type: hasKey ? "update" : "add",
                 object: this,
                 newValue: value,
                 name: key
             })
             if (!change) return this
-            value = change.newValue
+            value = change.newValue!
         }
         if (hasKey) {
             this._updateValue(key, value)
@@ -138,27 +126,24 @@ export class ObservableMap<V>
         return this
     }
 
-    delete(key: string): boolean {
-        this.assertValidKey(key)
-        key = "" + key
+    delete(key: K): boolean {
         if (hasInterceptors(this)) {
-            const change = interceptChange<IMapWillChange<V>>(this, {
+            const change = interceptChange<IMapWillChange<K, V>>(this, {
                 type: "delete",
                 object: this,
                 name: key
             })
             if (!change) return false
         }
-
         if (this._has(key)) {
             const notifySpy = isSpyEnabled()
             const notify = hasListeners(this)
             const change =
                 notify || notifySpy
-                    ? <IMapChange<V>>{
+                    ? <IMapChange<K, V>>{
                           type: "delete",
                           object: this,
-                          oldValue: (<any>this._data[key]).value,
+                          oldValue: (<any>this._data.get(key)).value,
                           name: key
                       }
                     : null
@@ -167,9 +152,9 @@ export class ObservableMap<V>
             transaction(() => {
                 this._keys.remove(key)
                 this._updateHasMapEntry(key, false)
-                const observable = this._data[key]!
+                const observable = this._data.get(key)!
                 observable.setNewValue(undefined as any)
-                this._data[key] = undefined as any
+                this._data.delete(key)
             })
             if (notify) notifyListeners(this, change)
             if (notifySpy) spyReportEnd()
@@ -178,31 +163,27 @@ export class ObservableMap<V>
         return false
     }
 
-    private _updateHasMapEntry(key: string, value: boolean): ObservableValue<boolean> {
+    private _updateHasMapEntry(key: K, value: boolean): ObservableValue<boolean> {
         // optimization; don't fill the hasMap if we are not observing, or remove entry if there are no observers anymore
-        let entry = this._hasMap[key]
+        let entry = this._hasMap.get(key)
         if (entry) {
             entry.setNewValue(value)
         } else {
-            entry = this._hasMap[key] = new ObservableValue(
-                value,
-                referenceEnhancer,
-                `${this.name}.${key}?`,
-                false
-            )
+            entry = new ObservableValue(value, referenceEnhancer, `${this.name}.${key}?`, false)
+            this._hasMap.set(key, entry)
         }
         return entry
     }
 
-    private _updateValue(key: string, newValue: V | undefined) {
-        const observable = this._data[key]!
+    private _updateValue(key: K, newValue: V | undefined) {
+        const observable = this._data.get(key)!
         newValue = (observable as any).prepareNewValue(newValue) as V
         if (newValue !== UNCHANGED) {
             const notifySpy = isSpyEnabled()
             const notify = hasListeners(this)
             const change =
                 notify || notifySpy
-                    ? <IMapChange<V>>{
+                    ? <IMapChange<K, V>>{
                           type: "update",
                           object: this,
                           oldValue: (observable as any).value,
@@ -210,7 +191,6 @@ export class ObservableMap<V>
                           newValue
                       }
                     : null
-
             if (notifySpy) spyReportStart({ ...change, name: this.name, key })
             observable.setNewValue(newValue as V)
             if (notify) notifyListeners(this, change)
@@ -218,39 +198,37 @@ export class ObservableMap<V>
         }
     }
 
-    private _addValue(key: string, newValue: V | undefined) {
+    private _addValue(key: K, newValue: V) {
         transaction(() => {
-            const observable = (this._data[key] = new ObservableValue(
+            const observable = new ObservableValue(
                 newValue,
                 this.enhancer,
                 `${this.name}.${key}`,
                 false
-            ))
+            )
+            this._data.set(key, observable)
             newValue = (observable as any).value // value might have been changed
             this._updateHasMapEntry(key, true)
             this._keys.push(key)
         })
-
         const notifySpy = isSpyEnabled()
         const notify = hasListeners(this)
         const change =
             notify || notifySpy
-                ? <IMapChange<V>>{
+                ? <IMapChange<K, V>>{
                       type: "add",
                       object: this,
                       name: key,
                       newValue
                   }
                 : null
-
         if (notifySpy) spyReportStart({ ...change, name: this.name, key })
         if (notify) notifyListeners(this, change)
         if (notifySpy) spyReportEnd()
     }
 
-    get(key: string): V | undefined {
-        key = "" + key
-        if (this.has(key)) return this.dehanceValue(this._data[key]!.get())
+    get(key: K): V | undefined {
+        if (this.has(key)) return this.dehanceValue(this._data.get(key)!.get())
         return this.dehanceValue(undefined)
     }
 
@@ -261,7 +239,7 @@ export class ObservableMap<V>
         return value
     }
 
-    keys(): Iterator<string> {
+    keys(): Iterator<K> {
         return (this._keys[iteratorSymbol()] as any)()
     }
 
@@ -277,7 +255,7 @@ export class ObservableMap<V>
         }
     }
 
-    entries(): Iterator<IMapEntry<V>> {
+    entries(): Iterator<IMapEntry<K, V>> {
         const self = this
         let nextIndex = 0
         return {
@@ -294,17 +272,18 @@ export class ObservableMap<V>
         }
     }
 
-    forEach(callback: (value: V, key: string, object: IMap<string, V>) => void, thisArg?) {
+    forEach(callback: (value: V, key: K, object: Map<K, V>) => void, thisArg?) {
         this._keys.forEach(key => callback.call(thisArg, this.get(key), key, this))
     }
 
     /** Merge another object into this object, returns this. */
-    merge(other: ObservableMap<V> | IKeyValueMap<V> | any): ObservableMap<V> {
+    merge(other: ObservableMap<K, V> | IKeyValueMap<V> | any): ObservableMap<K, V> {
         if (isObservableMap(other)) {
             other = other.toJS()
         }
         transaction(() => {
-            if (isPlainObject(other)) Object.keys(other).forEach(key => this.set(key, other[key]))
+            if (isPlainObject(other))
+                Object.keys(other).forEach(key => this.set((key as any) as K, other[key]))
             else if (Array.isArray(other)) other.forEach(([key, value]) => this.set(key, value))
             else if (isES6Map(other)) other.forEach((value, key) => this.set(key, value))
             else if (other !== null && other !== undefined)
@@ -321,17 +300,15 @@ export class ObservableMap<V>
         })
     }
 
-    replace(values: ObservableMap<V> | IKeyValueMap<V> | any): ObservableMap<V> {
+    replace(values: ObservableMap<K, V> | IKeyValueMap<V> | any): ObservableMap<K, V> {
         transaction(() => {
             // grab all the keys that are present in the new map but not present in the current map
             // and delete them from the map, then merge the new map
             // this will cause reactions only on changed values
-            const newKeys = getMapLikeKeys(values)
+            const newKeys = (getMapLikeKeys(values) as any) as K[]
             const oldKeys = this._keys
             const missingKeys = oldKeys.filter(k => newKeys.indexOf(k) === -1)
-
             missingKeys.forEach(k => this.delete(k))
-
             this.merge(values)
         })
         return this
@@ -342,32 +319,29 @@ export class ObservableMap<V>
     }
 
     /**
-     * Returns a shallow non observable object clone of this map.
-     * Note that the values might still be observable. For a deep clone use mobx.toJS.
-     */
-    toJS(): IKeyValueMap<V> {
+	 * Returns a plain object that represents this map.
+	 * Note that all the keys being stringified.
+	 * If there are duplicating keys after converting them to strings, behaviour is undetermined.
+	 */
+    toPOJO(): IKeyValueMap<V> {
         const res: IKeyValueMap<V> = {}
-        this._keys.forEach(key => (res[key] = this.get(key)!))
+        this._keys.forEach(key => (res["" + key] = this.get(key)!))
+        return res
+    }
+
+    /**
+	 * Returns a shallow non observable object clone of this map.
+	 * Note that the values migth still be observable. For a deep clone use mobx.toJS.
+	 */
+    toJS(): Map<K, V> {
+        const res: Map<K, V> = new Map()
+        this._keys.forEach(key => res.set(key, this.get(key)!))
         return res
     }
 
     toJSON(): IKeyValueMap<V> {
         // Used by JSON.stringify
-        return this.toJS()
-    }
-
-    private isValidKey(key: string) {
-        if (key === null || key === undefined) return false
-        if (typeof key === "string" || typeof key === "number" || typeof key === "boolean")
-            return true
-        return false
-    }
-
-    private assertValidKey(key: string) {
-        if (!this.isValidKey(key))
-            throw new Error(
-                `[mobx.map] Invalid key: '${key}', only strings, numbers and booleans are accepted as key in observable maps.`
-            )
+        return this.toPOJO()
     }
 
     toString(): string {
@@ -380,11 +354,11 @@ export class ObservableMap<V>
     }
 
     /**
-     * Observes this object. Triggers for the events 'add', 'update' and 'delete'.
-     * See: https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Object/observe
-     * for callback details
-     */
-    observe(listener: (changes: IMapChange<V>) => void, fireImmediately?: boolean): Lambda {
+	 * Observes this object. Triggers for the events 'add', 'update' and 'delete'.
+	 * See: https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Object/observe
+	 * for callback details
+	 */
+    observe(listener: (changes: IMapChange<K, V>) => void, fireImmediately?: boolean): Lambda {
         process.env.NODE_ENV !== "production" &&
             invariant(
                 fireImmediately !== true,
@@ -393,7 +367,7 @@ export class ObservableMap<V>
         return registerListener(this, listener)
     }
 
-    intercept(handler: IInterceptor<IMapWillChange<V>>): Lambda {
+    intercept(handler: IInterceptor<IMapWillChange<K, V>>): Lambda {
         return registerInterceptor(this, handler)
     }
 }
@@ -405,4 +379,4 @@ declareIterator(ObservableMap.prototype, function() {
 /* 'var' fixes small-build issue */
 export var isObservableMap = createInstanceofPredicate("ObservableMap", ObservableMap) as (
     thing: any
-) => thing is ObservableMap<any>
+) => thing is ObservableMap<any, any>